+++ docs/object-detection.md
@@ -19,11 +19,11 @@ Each command is configured with a config file in the form of a JSON-formatted [P
 
 Data is organized into a set of *training projects* and a set of *test projects*. Each training project consists of a *raster source* with imagery covering some area of interest, and a ground truth *annotation source* with labels generated by people. Each test project additionally contains a prediction annotation source where predictions are stored.
 
-### Machine learning methods and backends
+### Machine learning tasks and backends
 
-A *machine learning method* implements functionality specific to an abstract prediction problem such as object detection. In contrast, a *machine learning backend* implements a bridge to a (usually third party) concrete implementation of that method, such as the [Tensorflow Object Detection API](https://github.com/tensorflow/models/tree/master/research/object_detection).
+A *machine learning task* implements functionality specific to an abstract prediction problem such as object detection. In contrast, a *machine learning backend* implements a bridge to a (usually third party) concrete implementation of that task, such as the [Tensorflow Object Detection API](https://github.com/tensorflow/models/tree/master/research/object_detection).
 
-Raster Vision is designed to be easy to extend to new raster and annotation sources, and machine learning methods and backends.
+Raster Vision is designed to be easy to extend to new raster and annotation sources, and machine learning tasks and backends.
 
 ### Workflows
