+++ src/rv2/builders/predict_builder.py
@@ -1,5 +1,5 @@
 from rv2.builders import (
-    project_builder, ml_method_builder, label_map_builder)
+    project_builder, ml_task_builder, label_map_builder)
 from rv2.utils import files
 from rv2.commands.predict import Predict
 from rv2.protos.predict_pb2 import PredictConfig
@@ -9,10 +9,10 @@ def build(config):
     if isinstance(config, str):
         config = files.load_json_config(config, PredictConfig())
 
-    ml_method = ml_method_builder.build(config.machine_learning)
+    ml_task = ml_task_builder.build(config.machine_learning)
     label_map = label_map_builder.build(config.label_items)
     projects = [project_builder.build(project_config)
                 for project_config in config.projects]
     options = config.options
 
-    return Predict(projects, ml_method, label_map, options)
+    return Predict(projects, ml_task, label_map, options)
