+++ src/rv2/builders/train_builder.py
@@ -1,4 +1,4 @@
-from rv2.builders import ml_method_builder
+from rv2.builders import ml_task_builder
 from rv2.utils import files
 from rv2.commands.train import Train
 from rv2.protos.train_pb2 import TrainConfig
@@ -8,7 +8,7 @@ def build(config):
     if isinstance(config, str):
         config = files.load_json_config(config, TrainConfig())
 
-    ml_method = ml_method_builder.build(config.machine_learning)
+    ml_task = ml_task_builder.build(config.machine_learning)
     options = config.options
 
-    return Train(ml_method, options)
+    return Train(ml_task, options)
