+++ src/rv2/commands/train.py
@@ -2,9 +2,9 @@ from rv2.core.command import Command
 
 
 class Train(Command):
-    def __init__(self, ml_method, options):
-        self.ml_method = ml_method
+    def __init__(self, ml_task, options):
+        self.ml_task = ml_task
         self.options = options
 
     def run(self):
-        self.ml_method.train(self.options)
+        self.ml_task.train(self.options)
